blockTemplate can include transactions exceeding the max blocksize size
While populating the blockTemplate, it doesn't account for the
maxBlockSize which is the only thing that addBlock checks.  It is
possible to include a collection of transactions that would make the
block exceed this limit, making it unmineable and sticking the
blockchain.  The fix here ensures that the collection of tx in the
template does not exceed the current maxBlockSize.